refactor(#26): write -> filter-options endpoint path edit
