+++ api/src/main/java/co/kr/cocomu/study/controller/StudyController.java
@@ -9,9 +9,9 @@ import co.kr.cocomu.study.dto.response.AllStudyCardDto;
 import co.kr.cocomu.study.dto.response.LanguageDto;
 import co.kr.cocomu.study.dto.response.StudyCardDto;
 import co.kr.cocomu.study.dto.response.StudyDetailPageDto;
-import co.kr.cocomu.study.dto.response.StudyPageDto;
+import co.kr.cocomu.study.dto.page.StudyPageDto;
 import co.kr.cocomu.study.dto.response.WorkbookDto;
-import co.kr.cocomu.study.dto.response.WritePageDto;
+import co.kr.cocomu.study.dto.response.FilterOptionsDto;
 import co.kr.cocomu.study.service.StudyCommandService;
 import co.kr.cocomu.study.service.StudyQueryService;
 import jakarta.validation.Valid;
@@ -36,11 +36,11 @@ public class StudyController implements StudyControllerDocs {
     private final StudyCommandService studyCommandService;
     private final StudyQueryService studyQueryService;
 
-    @GetMapping("/write")
-    public Api<WritePageDto> getWritePageInfo(@AuthenticationPrincipal final Long userId) {
+    @GetMapping("/filter-options")
+    public Api<FilterOptionsDto> getWritePageInfo(@AuthenticationPrincipal final Long userId) {
         final List<LanguageDto> allLanguages = studyQueryService.getAllLanguages();
         final List<WorkbookDto> allWorkbooks = studyQueryService.getAllWorkbooks();
-        final WritePageDto result = new WritePageDto(allWorkbooks, allLanguages);
+        final FilterOptionsDto result = new FilterOptionsDto(allWorkbooks, allLanguages);
 
         return Api.of(StudyApiCode.GET_WRITE_PAGE_SUCCESS, result);
     }
