+++ api/src/main/java/co/kr/cocomu/study/controller/docs/StudyControllerDocs.java
@@ -7,15 +7,13 @@ import co.kr.cocomu.study.dto.request.GetAllStudyFilterDto;
 import co.kr.cocomu.study.dto.response.AllStudyCardDto;
 import co.kr.cocomu.study.dto.response.StudyCardDto;
 import co.kr.cocomu.study.dto.response.StudyDetailPageDto;
-import co.kr.cocomu.study.dto.response.StudyPageDto;
-import co.kr.cocomu.study.dto.response.WritePageDto;
+import co.kr.cocomu.study.dto.page.StudyPageDto;
+import co.kr.cocomu.study.dto.response.FilterOptionsDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.PathVariable;
 
 @Tag(name = "003. COCOMU-STUDY", description = "코코무 스터디 관련 API")
 public interface StudyControllerDocs {
@@ -97,5 +95,5 @@ public interface StudyControllerDocs {
         responseCode = "200",
         description = "스터디 작성 페이지 조회에 성공했습니다."
     )
-    Api<WritePageDto> getWritePageInfo(Long userId);
+    Api<FilterOptionsDto> getWritePageInfo(Long userId);
 }
