+++ api/src/main/java/co/kr/cocomu/study/dto/response/FilterOptionsDto.java
@@ -3,8 +3,8 @@ package co.kr.cocomu.study.dto.response;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 
-@Schema(description = "스터디 생성 페이지 응답")
-public record WritePageDto(
+@Schema(description = "스터디 필터링 옵션 응답")
+public record FilterOptionsDto(
     @Schema(description = "스터디 전체 문제집 정보")
     List<WorkbookDto> workbooks,
     @Schema(description = "스터디 전체 언어 정보")
